refactor: try-with-resouce 구문으로 변경
