+++ app/src/main/java/com/techcourse/dao/QueryExecutor.java
@@ -1,7 +1,10 @@
-package com.techcourse.dao.executor;
+package com.techcourse.dao;
 
 import com.interface21.jdbc.core.JdbcTemplate;
 import com.interface21.jdbc.core.Parameters;
+import com.interface21.jdbc.core.RowMapper;
+
+import java.util.List;
 
 public class QueryExecutor {
 
@@ -16,9 +19,15 @@ public class QueryExecutor {
         Parameters createParameters();
     }
 
-    public void execute(String sql, ParameterSetter parameterSetter) {
+    public void update(String sql, ParameterSetter parameterSetter) {
         final var parameters = parameterSetter.createParameters();
 
         jdbcTemplate.update(sql, parameters);
+    }
+
+    public <T> List<T> query(String sql, ParameterSetter parameterSetter, RowMapper<T> rowMapper) {
+        final var parameters = parameterSetter.createParameters();
+
+        return jdbcTemplate.query(sql, parameters, rowMapper);
     }
 }
